Implement logic to upload payload to endpoint
